// src/src/lib/database.types.ts
export interface Database {
  public: {
    Tables: {
// ...
      customers: {
        Row: {
          id: string
          name: string
          logo_url: string | null
          created_at: string
          slug: string | null
          category: string | null
        }
        Insert: {
          id?: string
          name: string
          logo_url?: string | null
          created_at?: string
          slug?: string | null
          category?: string | null
        }
        Update: {
          name?: string
          logo_url?: string | null
          slug?: string | null
          category?: string | null
        }
        Relationships: []
      }
// ...
    }

    Views: Record<string, never>
    Functions: Record<string, never>
    Enums: Record<string, never>
    CompositeTypes: Record<string, never>
  }
}

export type ProfileRow = Database['public']['Tables']['profiles']['Row']
export type CategoryRow = Database['public']['Tables']['categories']['Row']
export type ProductRow = Database['public']['Tables']['products']['Row']
export type CustomerRow = Database['public']['Tables']['customers']['Row']